+++ src/application/structs/UnlockStatus.h
@@ -3,7 +3,14 @@
 struct UnlockStatus
 {
     int currentEmblems;
-    // int emblemGoal;
+
+    bool whiteEmerald;
+    bool redEmerald;
+    bool cyanEmerald;
+    bool purpleEmerald;
+    bool greenEmerald;
+    bool yellowEmerald;
+    bool blueEmerald;
 
     bool sonicUnlocked;
     bool sonicLightShoesUnlocked;
@@ -30,7 +37,7 @@ struct UnlockStatus
     bool bigLifeRingUnlocked;
     bool bigPowerRodUnlocked;
     int bigLureQuantity = 0;
-    
+
     bool keyTrain;
     bool keyBoat;
     bool keyRaft;
@@ -43,4 +50,9 @@ struct UnlockStatus
     bool keyIceStone;
     bool keyWindStone;
     bool jungleCart;
+
+    bool GotAllChaosEmeralds()
+    {
+        return whiteEmerald && redEmerald && cyanEmerald && purpleEmerald && greenEmerald && yellowEmerald && blueEmerald;
+    }
 };